% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_products.R
\name{filter_products}
\alias{filter_products}
\title{Retrieve structured products data from thousands of online retailers and e-commerce sites}
\usage{
filter_products(query, sort = "relevancy", ts = (Sys.time() - (3 * 24 * 60 *
  60)), size = 100, from = 0, quiet = !interactive(),
  token = Sys.getenv("WEBHOSE_TOKEN"), ...)
}
\arguments{
\item{query}{A string query containing the filters that define which products will be returned.}

\item{sort}{By default the results are sorted by relevancy. Acceptable values are
"\code{relevancy}", "\code{social.facebook.likes}", "\code{social.facebook.shares}",
"\code{social.facebook.comments}", "\code{social.gplus.shares}", "\code{social.pinterest.shares}",
"\code{social.linkedin.shares}", "\code{social.stumbledupon.shares}", "\code{social.vk.shares}",
"\code{replies_count}", "\code{participants_count}", "\code{spam_score}", "\code{performance_score}",
"\code{published}", "\code{thread.published}", "\code{domain_rank}", "\code{ord_in_thread}",
"\code{rating}".}

\item{ts}{A timestamp to start the search from. If a \code{POSIXct} is passed in, it will
be converted to the necessary value in milliseconds. Default is previous 3 days.}

\item{size}{Total number of posts returned per request, ranges between \code{1:100}
(default is \code{100}).}

\item{from}{Paging parameter (starting record number). Default is \code{0}.}

\item{quiet}{By default, calls in interactive sessions will return updates during fetching.
Use \code{TRUE} to suppress these messages.}

\item{token}{Your private access token. You get a unique access token when you sign up.
Store it in an environment variable \code{WEBHOSE_TOKEN} (usually in \code{~/.Renviron})
or provide it directly.}

\item{...}{other parameters passed on to \code{httr::GET()}}
}
\value{
a \code{list} with these fields:
\itemize{
\item \code{totalResults}: The total number of posts matching your query (numeric)
\item \code{moreResultsAvailable}: How many more results are available (numeric)
\item \code{next}: A URL to get the next batch of posts matching your query. (character)
\item \code{requestsLeft}: How many more requests are available in your current subscription plan. (numeric)
}
}
\description{
Retrieve structured products data from thousands of online retailers and e-commerce sites
}
\examples{
\dontrun{
res <- filter_products("name:iphone")
}
}
\references{
\href{https://docs.webhose.io/docs/get-parameters}{webhose API}
}
