% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_reviews.R
\name{filter_reviews}
\alias{filter_reviews}
\title{Retrieve structured reviews data from hundreds of review sites}
\usage{
filter_reviews(query, sort = "relevancy", ts = (Sys.time() - (3 * 24 * 60 *
  60)), order = "desc", size = 100, accuracy_confidence = NULL,
  highlight = FALSE, from = 0, quiet = !interactive(),
  token = Sys.getenv("WEBHOSE_TOKEN"), ...)
}
\arguments{
\item{query}{A string query containing the filters that define which posts will be returned.}

\item{sort}{By default the results are sorted by relevancy. Acceptable values are
"\code{relevancy}", "\code{social.facebook.likes}", "\code{social.facebook.shares}",
"\code{social.facebook.comments}", "\code{social.gplus.shares}", "\code{social.pinterest.shares}",
"\code{social.linkedin.shares}", "\code{social.stumbledupon.shares}", "\code{social.vk.shares}",
"\code{replies_count}", "\code{participants_count}", "\code{spam_score}", "\code{performance_score}",
"\code{published}", "\code{thread.published}", "\code{domain_rank}", "\code{ord_in_thread}",
"\code{rating}".}

\item{ts}{A timestamp to start the search from. If a \code{POSIXct} is passed in, it will
be converted to the necessary value in milliseconds. Default is previous 3 days.}

\item{order}{\code{asc} (ascending) or \code{desc} (descending, default) sort order for results}

\item{size}{Total number of posts returned per request, ranges between \code{1:100}
(default is \code{100}).}

\item{accuracy_confidence}{\code{NULL} or \code{high}. If \code{high}, return only posts with high
extraction accuracy, but removes about 30% of the total matching posts (with
lower confidence).}

\item{highlight}{\code{FALSE} or \code{TRUE}. Return the fragments in the post that matched the
textual boolean query. The matched keywords will be surrounded by \code{<em/>} tags.
Default: \code{FALSE}}

\item{from}{Paging parameter (starting record number). Default is \code{0}.}

\item{quiet}{By default, calls in interactive sessions will return updates during fetching.
Use \code{TRUE} to suppress these messages.}

\item{token}{Your private access token. You get a unique access token when you sign up.
Store it in an environment variable \code{WEBHOSE_TOKEN} (usually in \code{~/.Renviron})
or provide it directly.}

\item{...}{other parameters passed on to \code{httr::GET()}}
}
\value{
a \code{list} with these fields:
\itemize{
\item \code{totalResults}: The total number of posts matching your query (numeric)
\item \code{moreResultsAvailable}: How many more results are available (numeric)
\item \code{next}: A URL to get the next batch of posts matching your query. (character)
\item \code{requestsLeft}: How many more requests are available in your current subscription plan. (numeric)
}
}
\description{
Retrieve structured reviews data from hundreds of review sites
}
\examples{
\dontrun{
res <- filter_reviews("site_category:travel rating:<3 (bug OR roach OR cockroach)")
}
}
\references{
\href{https://docs.webhose.io/docs/get-parameters}{webhose API}
}
