% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhose-package.R
\docType{package}
\name{webhose}
\alias{webhose}
\alias{webhose-package}
\title{Tools to Work with the 'webhose.io' 'API'}
\description{
The 'webhose.io' <https://webhose.io/about> 'API' provides access to structured
web data feeds across vertical content domains. Their crawlers download the web,
structure the data and index save it into domain-specific repositories that can be
accessed on demand. Methods are provided to query and retrieve content from this 'API'.
}
\author{
Bob Rudis (bob@rud.is)
}
